requirements.txt
discord.py
notion-client
python-dotenv

